Pluralist governments for Cosmopolitan religions
Pluralist Feudal is similar to Feudal, without wrong religion maluses
and the possibility to demand conversion (similar to Secular vs Pious
Feudal). It is enabled for Temporal RHs and rulers crowned by a temporal
or Pope-tier theocratic RH of their current religion.

Pluralist Imperial is based on Chinese Imperial. It is available to
rulers of empire with the pluralist_imperial flag, but the method to
acquire this flag is still WIP.